Refactor Renderer component to accept CharRow or Zipper instead of Scene
