+++ demo/src/solver/substeps.tsx
@@ -3,8 +3,7 @@ import * as React from "react";
 import * as Editor from "@math-blocks/editor";
 import * as Semantic from "@math-blocks/semantic";
 import * as Solver from "@math-blocks/solver";
-import * as Typesetter from "@math-blocks/typesetter";
-import {MathRenderer, FontDataContext} from "@math-blocks/react";
+import {MathRenderer} from "@math-blocks/react";
 
 type Props = {
     // Prefix to start numbering from, e.g. 1.2.3
@@ -17,32 +16,20 @@ type Props = {
 };
 
 const Substeps: React.FunctionComponent<Props> = ({prefix, start, step}) => {
-    const fontData = React.useContext(FontDataContext);
-
     let current = start;
 
-    const fontSize = 64;
-    const context: Typesetter.Context = {
-        fontData,
-        baseFontSize: fontSize,
-        mathStyle: Typesetter.MathStyle.Display,
-        renderMode: Typesetter.RenderMode.Static,
-        cramped: false,
-        // colorMap: props.colorMap,
-    };
-
     const beforeRow = Editor.print(step.before);
-    const beforeScene = Typesetter.typeset(beforeRow, context);
+    console.log(beforeRow);
 
     return (
         <div style={{display: "flex", flexDirection: "column"}}>
-            <MathRenderer scene={beforeScene} style={{marginBottom: 32}} />
+            <MathRenderer row={beforeRow} style={{marginBottom: 32}} />
             {step.substeps.map((substep, index) => {
                 const before = current;
 
                 const after = Solver.applyStep(before, substep);
                 const afterRow = Editor.print(substep.after);
-                const afterScene = Typesetter.typeset(afterRow, context);
+                console.log(afterRow);
 
                 current = after;
 
@@ -65,7 +52,7 @@ const Substeps: React.FunctionComponent<Props> = ({prefix, start, step}) => {
                         )}
                         {
                             <MathRenderer
-                                scene={afterScene}
+                                row={afterRow}
                                 style={{marginBottom: 32}}
                             />
                         }
