+++ packages/react/src/math-renderer.tsx
@@ -1,182 +1,80 @@
 import * as React from "react";
 
-import {UnreachableCaseError} from "@math-blocks/core";
-import {SceneGraph} from "@math-blocks/typesetter";
-import * as OpenType from "@math-blocks/opentype";
+import * as Editor from "@math-blocks/editor";
+import * as Typesetter from "@math-blocks/typesetter";
 
-const Line: React.FunctionComponent<SceneGraph.Line> = ({
-    id,
-    style,
-    thickness,
-    ...props
-}) => {
-    return (
-        <line
-            {...props}
-            stroke={style.stroke}
-            strokeWidth={thickness}
-            strokeLinecap="butt"
-        />
-    );
-};
+import {FontDataContext} from "./font-data-context";
+import SceneRenderer from "./scene-renderer";
 
-const Rect: React.FunctionComponent<SceneGraph.Rect> = ({
-    fill,
-    id,
-    ...props
-}) => {
-    return <rect {...props} fill={fill} />;
-};
-
-const getPath = (glyph: OpenType.Glyph): string => {
-    let result = "";
-
-    // The glyph's path is in font units.
-    const path = glyph.path;
-
-    for (const cmd of path) {
-        if (cmd.type === "M") {
-            result += `M ${cmd.x},${cmd.y} `;
-        } else if (cmd.type === "L") {
-            result += `L ${cmd.x},${cmd.y} `;
-        } else if (cmd.type === "C") {
-            result += `C ${cmd.x1},${cmd.y1} ${cmd.x2},${cmd.y2} ${cmd.x},${cmd.y}`;
-        } else if (cmd.type === "Q") {
-            result += `Q ${cmd.x1},${cmd.y1} ${cmd.x},${cmd.y}`;
-        } else {
-            result += "Z";
-        }
-    }
-
-    return result;
-};
-
-enum GlyphRendering {
-    Path,
-    Text,
-}
-
-const Glyph: React.FunctionComponent<SceneGraph.Glyph> = ({
-    x,
-    y,
-    glyph,
-    style,
-}) => {
-    const id = typeof glyph.id !== undefined ? String(glyph.id) : undefined;
-
-    const {font} = glyph.fontData;
-    const scale = glyph.size / font.head.unitsPerEm;
-
-    const glyphRendering: GlyphRendering = GlyphRendering.Path;
-
-    if (glyphRendering === GlyphRendering.Path) {
-        return (
-            <path
-                fill={style.fill}
-                id={id}
-                style={{opacity: glyph.pending ? 0.5 : 1.0}}
-                aria-hidden="true"
-                d={getPath(font.getGlyph(glyph.glyphID))}
-                transform={`translate(${x}, ${y}) scale(${scale}, -${scale})`}
-            />
-        );
-    } else {
-        const {fontFamily} = glyph.fontData;
-        return (
-            <text
-                x={x}
-                y={y}
-                fontFamily={fontFamily}
-                fontSize={glyph.size}
-                fill={style.fill}
-                id={id}
-                style={{opacity: glyph.pending ? 0.5 : 1.0}}
-                aria-hidden="true"
-            >
-                {glyph.char}
-            </text>
-        );
-    }
-};
-
-const Group: React.FunctionComponent<SceneGraph.Group> = ({
-    x,
-    y,
-    children,
-    style,
-    id,
-}) => {
-    const _id = typeof id !== undefined ? String(id) : undefined;
-
-    return (
-        <g
-            transform={`translate(${x},${y})`}
-            fill={style.fill}
-            stroke={style.stroke}
-            id={_id}
-        >
-            {children.map((child, i) => {
-                const key = `${i}`;
-                return <Node {...child} key={key} />;
-            })}
-        </g>
-    );
-};
-
-const Node: React.FunctionComponent<SceneGraph.Node> = (props) => {
-    switch (props.type) {
-        case "char":
-            return <Glyph {...props} />;
-        case "group":
-            return <Group {...props} />;
-        case "line":
-            return <Line {...props} />;
-        case "rect":
-            return <Rect {...props} />;
-        default:
-            throw new UnreachableCaseError(props);
-    }
-};
-
-const CURSOR_WIDTH = 2;
+const {useContext} = React;
 
 type Props = {
-    readonly scene: SceneGraph.Scene;
+    // The initial value for the editor
+    readonly row?: Editor.types.CharRow;
+    readonly zipper?: Editor.Zipper;
+    readonly fontSize?: number;
+    readonly showCursor?: boolean;
+    readonly radicalDegreeAlgorithm?: Typesetter.RadicalDegreeAlgorithm;
+    readonly mathStyle?: Typesetter.MathStyle;
+    readonly renderMode?: Typesetter.RenderMode;
+
+    // Style
     readonly style?: React.CSSProperties;
+
+    // Renders bounding boxes around each group and glyph.
     readonly showHitboxes?: boolean;
 };
 
-const MathRenderer = React.forwardRef<SVGSVGElement, Props>((props, ref) => {
-    const {scene, style, showHitboxes} = props;
-    const {width, height} = scene;
-    const padding = CURSOR_WIDTH / 2;
-    const viewBox = `-${padding} 0 ${width + CURSOR_WIDTH} ${height}`;
+// TODO: expose other settings such as display style as props
+// TODO: add an onBlur prop
+export const MathRenderer = React.forwardRef<SVGSVGElement, Props>(
+    (props, ref) => {
+        const fontData = useContext(FontDataContext);
+        const {style, fontSize, showHitboxes} = props;
 
-    return (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox={viewBox}
-            width={width + CURSOR_WIDTH}
-            height={height}
-            style={style}
-            ref={ref}
-        >
-            <g fill="currentColor" stroke="currentColor">
-                <Group {...scene.selection} />
-                {/**
-                 * We set 'fill' and stroke to 'currentColor' so that the base
-                 * color is whatever the current CSS 'color' property is set up.
-                 * Individual nodes within the scene can override their style's
-                 * color and the render above will set the fill and/or stroke in
-                 * rendered SVG element appropriately.
-                 */}
-                <Group {...scene.content} />
-            </g>
-            {showHitboxes && <Group {...scene.hitboxes} />}
-        </svg>
-    );
-});
+        const context: Typesetter.Context = {
+            fontData: fontData,
+            baseFontSize: fontSize || 64,
+            mathStyle: Typesetter.MathStyle.Display,
+            cramped: false,
+            renderMode: Typesetter.RenderMode.Dynamic,
+            radicalDegreeAlgorithm: props.radicalDegreeAlgorithm,
+        };
+
+        const options = {showCursor: props.showCursor, debug: true};
+
+        const scene = (() => {
+            if (props.row) {
+                return Typesetter.typeset(props.row, context, options);
+            } else if (props.zipper) {
+                return Typesetter.typesetZipper(props.zipper, context, options);
+            } else {
+                return null;
+            }
+        })();
+
+        if (scene == null) {
+            return null;
+        }
+
+        return (
+            <SceneRenderer
+                ref={ref}
+                scene={scene}
+                showHitboxes={showHitboxes}
+                style={style}
+            />
+        );
+    },
+);
 
 MathRenderer.displayName = "MathRenderer";
+
+MathRenderer.defaultProps = {
+    style: {},
+    fontSize: 64,
+    renderMode: Typesetter.RenderMode.Static,
+    mathStyle: Typesetter.MathStyle.Display,
+};
 
 export default MathRenderer;
